api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
supabase==2.3.4
python-dotenv==1.0.1
